requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.2.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
click-log==0.4.0
Crawl4AI==0.8.5
cryptography==46.0.5
cssselect==1.4.0
dataclasses-json==0.6.7
ddgs==9.11.4
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
fake-useragent==2.2.0
fastapi==0.135.1
fastuuid==0.14.0
filelock==3.25.2
firecrawl-py==4.20.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==0.37.1
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-groq==1.1.2
langchain-ollama==1.0.1
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lark==1.3.1
litellm==1.82.4
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.4
openai==2.29.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
patchright==1.58.2
pillow==12.1.1
playwright==1.58.0
playwright-stealth==2.0.2
primp==1.1.3
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pyOpenSSL==26.0.0
pypdf==6.9.1
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rtree==1.4.1
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
shiboken6==6.10.2
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
trimesh==4.11.4
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
yt-dlp==2026.3.17
zipp==3.23.0
zstandard==0.25.0
